Split generic part of generate_unifrac_task_noclass into separate ssu_generate_helper module
